Update the inverter limit whenever appropriate
The retention time for the next update is shorter the more the new power
limit differs from the last one. The next update is calculated by

    time_passed > (200 / sqrt(abs(current_limit - new_limit) + 1)

That means that if the power consumption changes by 1000 W, the minimum
time between updates is ~ 6.3 sec, whereas for a difference of 100 W the
time is ~ 20 sec and for 10 W the ~ 60 sec.